+++ CubeIDE/Code/inc/main.h
@@ -23,14 +23,14 @@ struct main_flags_struct
 {
     uint8_t buttons_scanned;
     uint8_t parse_nmea;
+    uint8_t start_radio;
     uint8_t update_screen;
-    uint8_t run_frame;
     uint8_t do_beep;
 
     uint8_t tx_state;
     uint8_t rx_state;
 
-    uint8_t frame_ended;
+    uint8_t process_all;
     uint8_t nmea_parsed_only;
     uint8_t pps_synced;