+++ CubeIDE/Code/inc/radio.h
@@ -14,6 +14,7 @@
 void rf_init(void);
 uint8_t rf_tx_packet(void);
 uint8_t rf_start_rx(void);
+void rf_workaround_15_3(void);
 void rf_get_rx_packet(void);
 void rf_set_standby_xosc(void);
 uint8_t rf_get_status(void);