+++ CubeIDE/Code/inc/service.h
@@ -28,7 +28,7 @@ float atof32(char *input);
 void ftoa32(float value, uint8_t precision, char *buffer);
 int32_t atoi32(char *input);
 void itoa32(int32_t value, char *buffer);
-void add_leading_zero(char *buf);
+void time_date_add_leading_zero(char *buf);