+++ CubeIDE/Code/src/lcd.c
@@ -86,7 +86,7 @@ void lcd_init(void)
         cs_lcd_active();
         while (len--)
         {
-            spi2_trx(sh1106_conf[i++]);
+            spi2_tx(sh1106_conf[i++]);
         }
         cs_lcd_inactive();
     }
@@ -105,7 +105,7 @@ void lcd_send_command(uint8_t command)
 	//spi2_clock_enable();
 	lcd_command_mode();
 	cs_lcd_active();
-	spi2_trx(command);
+	spi2_tx(command);
 	cs_lcd_inactive();
 	//spi2_clock_disable();
 }
@@ -159,7 +159,6 @@ void lcd_update(void)
 			//spi2_clock_enable();
 
 
-			//todo: if current_page <> 0 to prevent glitches
 			lcd_busy = 1;
 			current_page = 0;
 			lcd_send_command(0x02); 		//reset column address low to 2 because LCD panel is centered to SH1106 frame buffer
@@ -230,7 +229,7 @@ void lcd_pixel(uint8_t x, uint8_t y, int8_t action)
     switch (action)
     {
         case 0:     //clear pixel
-            screen_buf[x + (y / 8) * LCD_SIZE_X] &= ~(1 << (y % 8));		////todo: use bit band instead
+            screen_buf[x + (y / 8) * LCD_SIZE_X] &= ~(1 << (y % 8));
             break;
         case 1:     //set pixel
             screen_buf[x + (y / 8) * LCD_SIZE_X] |= 1 << (y % 8);
@@ -248,7 +247,7 @@ void lcd_pixel(uint8_t x, uint8_t y, int8_t action)
 void lcd_set_pixel(uint8_t x, uint8_t y)
 {
     //set pixel
-    screen_buf[x + (y / 8) * LCD_SIZE_X] |= 1 << (y % 8);			//todo: use bit band instead
+    screen_buf[x + (y / 8) * LCD_SIZE_X] |= 1 << (y % 8);
 }
 
 
@@ -259,7 +258,7 @@ void lcd_set_pixel_plot(uint8_t x, uint8_t y)
 {
     //set pixel
 	y = 63 - y;
-    screen_buf[x + (y / 8) * LCD_SIZE_X] |= 1 << (y % 8);			//todo: use bit band instead
+    screen_buf[x + (y / 8) * LCD_SIZE_X] |= 1 << (y % 8);
 }
 
 
@@ -273,33 +272,24 @@ void lcd_pos(uint8_t row, uint8_t col)
 
 
 //Put one char in buffer in position, defined previously via ssd1306_pos()
-void lcd_char(char chr, uint8_t inv)
+void lcd_char(char chr)
 {
 	uint16_t buf_pos_copy = buf_pos;
 	uint16_t font_char_pos, c;
 	uint8_t px;
-
-    if (inv)
-    {
-        inv = 0xFF;
-    }
-    else
-    {
-        inv = 0x00;
-    }
 
     //font_char_pos = ((uint8_t)chr - 32) * FONT_BYTES;
     font_char_pos = ((uint8_t)chr) * FONT_BYTES;
 
     for (px = 0, c = font_char_pos; px < FONT_BYTES_X; px++, c += 2)		//upper symbol row
     {
-        screen_buf[buf_pos++] = font_8x16[c] ^ inv;
+        screen_buf[buf_pos++] = font_8x16[c];
     }
 
     buf_pos = buf_pos_copy + LCD_SIZE_X; //point to the lower symbol's row
     for (px = 0, c = font_char_pos + 1; px < FONT_BYTES_X; px++, c += 2)	//lower symbol row
     {
-        screen_buf[buf_pos++] = font_8x16[c] ^ inv;
+        screen_buf[buf_pos++] = font_8x16[c];
     }
 
     buf_pos = buf_pos_copy + FONT_BYTES_X;	//point to the next LCD char
@@ -308,32 +298,23 @@ void lcd_char(char chr, uint8_t inv)
 
 
 
-void lcd_char16(char chr, uint8_t inv)
+void lcd_char16(char chr)
 {
 	uint16_t buf_pos_copy = buf_pos;
 	uint16_t font_char_pos, c;
 	uint8_t px;
 
-    if (inv)			//todo: delete inv everywhere, it is useless
-    {
-        inv = 0xFF;
-    }
-    else
-    {
-        inv = 0x00;
-    }
-
     font_char_pos = ((uint8_t)chr - 48) * FONT16_BYTES;
 
     for (px = 0, c = font_char_pos; px < FONT16_BYTES_X; px++, c += 2)		//upper symbol row
     {
-        screen_buf[buf_pos++] = font_16x16[c] ^ inv;
+        screen_buf[buf_pos++] = font_16x16[c];
     }
 
     buf_pos = buf_pos_copy + LCD_SIZE_X; //point to the lower symbol's row
     for (px = 0, c = font_char_pos + 1; px < FONT16_BYTES_X; px++, c += 2)	//lower symbol row
     {
-        screen_buf[buf_pos++] = font_16x16[c] ^ inv;
+        screen_buf[buf_pos++] = font_16x16[c];
     }
 
     buf_pos = buf_pos_copy + FONT16_BYTES_X;	//point to the next LCD char
@@ -343,61 +324,61 @@ void lcd_char16(char chr, uint8_t inv)
 
 
 //Put one char in defined pos
-void lcd_char_pos(uint8_t row, uint8_t col, char chr, uint8_t inv)
+void lcd_char_pos(uint8_t row, uint8_t col, char chr)
 {
     lcd_pos(row, col);
-    lcd_char(chr, inv);
+    lcd_char(chr);
 }
 
 
 
 //Put one char in defined pos
-void lcd_char16_pos(uint8_t row, uint8_t col, char chr, uint8_t inv)
+void lcd_char16_pos(uint8_t row, uint8_t col, char chr)
 {
     lcd_pos(row, col);
-    lcd_char16(chr, inv);
+    lcd_char16(chr);
 }
 
 
 
 //Print string on screen (with position (rows 0-3, cols 0-15))
-void lcd_print(uint8_t row, uint8_t col, char *p_str, uint8_t inv)
+void lcd_print(uint8_t row, uint8_t col, char *p_str)
 {
     lcd_pos(row, col);
     
     while (*p_str)
     {
-        lcd_char(*p_str++, inv);
+        lcd_char(*p_str++);
     }
 }
 
 
 
 //Clear, print, update
-void lcd_print_only(uint8_t row, uint8_t col, char *p_str, uint8_t inv)
+void lcd_print_only(uint8_t row, uint8_t col, char *p_str)
 {
 	lcd_clear();
-	lcd_print(row, col, p_str, inv);
+	lcd_print(row, col, p_str);
 	lcd_update();
 }
 
 
 
 
-void lcd_print16(uint8_t row, uint8_t col, char *p_str, uint8_t inv)
+void lcd_print16(uint8_t row, uint8_t col, char *p_str)
 {
     lcd_pos(row, col);
 
     while (*p_str)
     {
-        lcd_char16(*p_str++, inv);
+        lcd_char16(*p_str++);
     }
 }
 
 
 
 //Print string on screen (with position) in viceversa direction (decrease collumn)
-void lcd_print_viceversa(uint8_t row, uint8_t col, char *p_str, uint8_t inv)
+void lcd_print_viceversa(uint8_t row, uint8_t col, char *p_str)
 {
     uint8_t symb_cntr = 0;
     
@@ -412,7 +393,7 @@ void lcd_print_viceversa(uint8_t row, uint8_t col, char *p_str, uint8_t inv)
     while (symb_cntr)
     {
         symb_cntr--;
-        lcd_char(*--p_str, inv);
+        lcd_char(*--p_str);
         buf_pos -= 2 * FONT_SIZE_X;         //minus two characters position
     }
 }
@@ -420,7 +401,7 @@ void lcd_print_viceversa(uint8_t row, uint8_t col, char *p_str, uint8_t inv)
 
 
 //Print string on screen (with position) in viceversa direction (decrease collumn)
-void lcd_print16_viceversa(uint8_t row, uint8_t col, char *p_str, uint8_t inv)
+void lcd_print16_viceversa(uint8_t row, uint8_t col, char *p_str)
 {
     uint8_t symb_cntr = 0;
 
@@ -435,7 +416,7 @@ void lcd_print16_viceversa(uint8_t row, uint8_t col, char *p_str, uint8_t inv)
     while (symb_cntr)
     {
         symb_cntr--;
-        lcd_char16(*--p_str, inv);
+        lcd_char16(*--p_str);
         buf_pos -= 4 * FONT_SIZE_X;         //minus two characters position
     }
 }
@@ -443,17 +424,17 @@ void lcd_print16_viceversa(uint8_t row, uint8_t col, char *p_str, uint8_t inv)
 
 
 //Print string on screen
-void lcd_print_next(char *p_str, uint8_t inv)
+void lcd_print_next(char *p_str)
 {
     while (*p_str)
     {
-        lcd_char(*p_str++, inv);
+        lcd_char(*p_str++);
     }
 }
 
 
 
-void lcd_print_next_viceversa(char *p_str, uint8_t inv)	//use only after lcd_print_viceversa()
+void lcd_print_next_viceversa(char *p_str)	//use only after lcd_print_viceversa()
 {
     uint8_t symb_cntr = 0;
 
@@ -468,7 +449,7 @@ void lcd_print_next_viceversa(char *p_str, uint8_t inv)	//use only after lcd_pri
     while (symb_cntr)
     {
         symb_cntr--;
-        lcd_char(*--p_str, inv);
+        lcd_char(*--p_str);
         buf_pos -= 2 * FONT_SIZE_X;         //minus two characters position
     }
 }