+++ CubeIDE/Code/src/main.c
@@ -43,9 +43,11 @@ uint8_t processing_button = 0;
 //TIMERS
 uint32_t uptime = 0;
 uint32_t pps_absolute_counter = 0;
-uint32_t pps_relative_counter = 0;
-uint8_t time_slot_timer_ovf = 0;
-uint8_t time_slot = 0;
+
+uint8_t second_modulo = 0;
+uint8_t device_tx_second = 0;
+uint8_t max_rx_second = 0;
+
 
 uint8_t *p_update_interval_values;
 
@@ -65,43 +67,42 @@ int main(void)
     lcd_update();
     delay_cyc(400000);
 
-    lcd_print_only(0, 0, "uart..", 0);
-    uart1_init();
-    uart3_dma_init();
+    lcd_print_only(0, 0, "uart..");
+    uart_init();
 
-    lcd_print_only(0, 0, "settings..", 0);
+    lcd_print_only(0, 0, "settings..");
     settings_load();
 
-    lcd_print_only(0, 0, "timers..", 0);
+    lcd_print_only(0, 0, "timers..");
     timers_init();
 
-    lcd_print_only(0, 0, "adc..", 0);
+    lcd_print_only(0, 0, "adc..");
     adc_init();
     adc_start_bat_voltage_reading();
 
-    lcd_print_only(0, 0, "i2c..", 0);
+    lcd_print_only(0, 0, "i2c..");
     i2c_init();
 
-    lcd_print_only(0, 0, "radio..", 0);
+    lcd_print_only(0, 0, "radio..");
     rf_init();
 
-    lcd_print_only(0, 0, "core..", 0);
+    lcd_print_only(0, 0, "core..");
     init_lrns();
 
-    lcd_print_only(0, 0, "gps..", 0);
+    lcd_print_only(0, 0, "gps..");
     gps_init();
 
-    lcd_print_only(0, 0, "compass..", 0);
+    lcd_print_only(0, 0, "compass..");
     init_compass();
 
-    lcd_print_only(0, 0, "interrupts..", 0);
+    lcd_print_only(0, 0, "interrupts..");
     ext_int_init();
     enable_buttons_interrupts();
 
-    lcd_print_only(0, 0, "mem points..", 0);
+    lcd_print_only(0, 0, "mem points..");
     init_memory_points();
 
-    lcd_print_only(0, 0, "menu..", 0);
+    lcd_print_only(0, 0, "menu..");
     init_menu();
 
 
@@ -110,6 +111,10 @@ int main(void)
     p_settings = get_settings();
     p_gps_num = get_gps_num();
     p_update_interval_values = get_update_interval_values();
+
+    device_tx_second = (2 * (p_settings->device_number - 1));
+    max_rx_second = (2 * (p_settings->devices_on_air - 1));
+
 
 
     make_a_beep();	//end of device loading
@@ -135,15 +140,11 @@ int main(void)
 
             if (parse_gps() == 1)
             {
-				if 	(main_flags.pps_synced == 1)		//ready to txrx when pps exists, data is valid and current second divides by send interval without remainder
+				if 	(main_flags.pps_synced == 1)		//ready to txrx when pps exists & data is valid
 				{
 					if (p_gps_num->status == GPS_DATA_VALID)
 					{
-						if ((p_gps_num->second % p_update_interval_values[p_settings->update_interval_opt]) == 0)
-						{
-							fill_air_packet(uptime);
-							main_flags.run_frame = 1;
-						}
+						main_flags.start_radio = 1;
 					}
 				}
 				else
@@ -163,7 +164,7 @@ int main(void)
 
             if (is_battery_critical())
             {
-            	release_power();	//just turn off for now
+            	release_power();	//todo: just turn off for now
             }
 
             if (!(main_flags.pps_synced)) 	//when no PPS we still need timeout alarming once in a sec (mostly for our device to alarm about no PPS)
@@ -176,10 +177,10 @@ int main(void)
 
 
 
-        //Checks after receiving packets from all devices or after no txrx; performing beep
-        if (main_flags.frame_ended == 1)
+        //Checks after receiving a packet from a device; performing beep
+        if (main_flags.process_all == 1)
         {
-        	main_flags.frame_ended = 0;
+        	main_flags.process_all = 0;
         	process_all_devices();
         	calc_fence();
         	calc_timeout(uptime);
@@ -254,7 +255,6 @@ void DMA1_Channel3_IRQHandler(void)
     	make_a_long_beep();
     }
 
-    pps_relative_counter = 0;
     main_flags.pps_synced = 0;
     main_flags.parse_nmea = 1;
 
@@ -262,31 +262,26 @@ void DMA1_Channel3_IRQHandler(void)
 
 
 
-//GPS PPS interrupt
+//GPS PPS Interrupt
 void EXTI2_IRQHandler(void)
 {
 
-	EXTI->PR = EXTI_PR_PR2;		//clear interrupt
-	timer1_start();                 //the first thing to do is to start time slot timer right after PPS
+	EXTI->PR = EXTI_PR_PR2;			//clear interrupt
+	timer1_start_800ms();           //the first thing to do is to start gps acquire timer
 
-	uart3_dma_stop();				//fix the data
-	backup_and_clear_uart_buffer();
+	uart3_dma_stop();				//drop the previous data; there will be new after this PPS
+	clear_uart_buffer();
 	uart3_dma_restart();
 
 	pps_absolute_counter++;
-	pps_relative_counter++;
 
-	if (pps_relative_counter > PPS_SKIP) //skip first two pps impulses: skip first PPS - ignore previous nmea data; skip second PPS, but fix the nmea data acquired after first PPS
-	{
-		main_flags.pps_synced = 1;
-		main_flags.parse_nmea = 1;
-	}
+	main_flags.pps_synced = 1;
 
 }
 
 
 
-//Radio interrupt (PayloadReady or PacketSent)
+//Radio Interrupt (PayloadReady or PacketSent)
 void EXTI0_IRQHandler(void)
 {
     EXTI->PR = EXTI_PR_PR0;         //clear interrupt
@@ -297,101 +292,92 @@ void EXTI0_IRQHandler(void)
 	if (current_radio_status & IRQ_RX_DONE)	//Packet received
 	{
 		main_flags.rx_state = 0;
+		led_green_off();
+
+		rf_workaround_15_3();	//run if rx timeout was used
 
 		if (!(current_radio_status & IRQ_CRC_ERROR))	// if no CRC error
 		{
 			rf_get_rx_packet();
 			parse_air_packet(uptime);   //parse air data from another device (which has ended TX in the current time_slot)
-
-			if (get_current_device() == time_slot)
-			{
-				led_green_on(); //indicate successful rx event only if received device is active in menu, it will be switched off at the next timeslot interrupt
-			}
 		}
 	}
 	else if (current_radio_status & IRQ_TX_DONE)		//Packet transmission completed
 	{
 		main_flags.tx_state = 0;
-		led_green_on(); //indicate successful tx event, led will be switched off at the next timeslot interrupt
+		led_green_off();
 	}
 	else if (current_radio_status & IRQ_RX_TX_TIMEOUT)	//RX timeout only, because TX timeout feature is not used at all
 	{
 		main_flags.rx_state = 0;
+		led_green_off();
 		rf_set_standby_xosc();	//after RX TO it goes to Standby RC mode only (https://forum.lora-developers.semtech.com/t/sx1268-is-it-possible-to-configure-transition-to-stdby-xosc-after-cad-done-rx-timeout/1282)
 	}
-
-
 }
 
 
 
-								//	each byte is timer interrupt mask. timeslot_pattern[0] is the initial state and always zero
-								//	0 - skip interrupt
-								//	1 - do interrupt
-								//	2 - stop timer
-
-								//		|NMEA	|Slot 1		|Slot 2		|Slot 3		|Slot 4		|Slot 5		|Processing	|
-								//		|-------|-----------|-----------|-----------|-----------|-----------|-----------|
-								//   	0	50	100	150	200	250	300	350	400	450	500	550	600	650	700	750	800	850	900	950	1000 ms
-const uint8_t timeslot_pattern[] = {	0, 	0, 	1,	0,	0,	1,	0,	0,	1,	0,	0,	1,	0,	0,	1,	0,	0,	1,	0,	0,	0		};
-
-//Time slot interrupt
+//Timer1 GPS Acquire + Radio Run
 void TIM1_UP_IRQHandler(void)
 {
     TIM1->SR &= ~TIM_SR_UIF;                    //clear interrupt
-    led_green_off();		//switch off green led after successful tx/rx event
+    timer1_stop_reload();						//stop this timer
 
-    time_slot_timer_ovf++;             			//increment ovf counter
-
-    if (timeslot_pattern[time_slot_timer_ovf] == 1)
+    if (timer1_get_intrvl_type() == 1) //long interval 800 ms ended
     {
 
-    	time_slot++;
+    	//parse NMEA, run short timer
+    	timer1_start_100ms();
 
-        if (main_flags.run_frame == 1)
-        {
-    		if (time_slot >= (p_settings->devices_on_air + 1)) //devices_on_air regulates how many time slots are active
+		uart3_dma_stop();					//fix the nmea data after last pps (800 ms ago)
+		backup_and_clear_uart_buffer();
+		uart3_dma_restart();				//restart for a case when no next PPS occur, then dma ovf will trigger
+										//otherwise next PPS will reset dma again before the upcoming nmea data
+
+    	main_flags.parse_nmea = 1;			//ask to parse
+
+    }
+    else if (timer1_get_intrvl_type() == 2) //short interval 100 ms ended, resulting 800 ms + 100 ms=900 ms after last pps
+    {
+
+    	//important: NMEA must be parsed before execution of this code
+    	//start radio, run processing of all devices
+
+    	if (main_flags.start_radio == 1)
+    	{
+    		main_flags.start_radio = 0;
+
+    		//calc a remainder of current second division by update interval
+    		second_modulo = p_gps_num->second % p_update_interval_values[p_settings->update_interval_opt];
+
+    		//timeslots are at X0, X2, X4, X6, X8 second; where X is 0, 1, 2, 3, 4, 5 depending on the update interval
+    		if (second_modulo == device_tx_second)
     		{
-    			timer1_stop_reload();			//end of the last time slot
-    			time_slot_timer_ovf = 0;
-    			time_slot = 0;
-    			main_flags.run_frame = 0;
-    			main_flags.frame_ended = 1;
+    			//tx
+				fill_air_packet(uptime);
+				if (rf_tx_packet())
+				{
+					main_flags.tx_state = 1;
+					led_green_on();
+				}
     		}
-    		else
+    		else if ((second_modulo <= max_rx_second)  && ((second_modulo % 2) == 0))
     		{
-    			if (time_slot == p_settings->device_number)
-    			{
-    				if (rf_tx_packet())
-    				{
-    					main_flags.tx_state = 1;
-    				}
-    			}
-    			else
-    			{
-    				if (rf_start_rx())
-    				{
-    					main_flags.rx_state = 1;
-    				}
-    			}
+    			//rx
+				if (rf_start_rx())
+				{
+					main_flags.rx_state = 1;
+
+					if (second_modulo == (2 * (get_current_device() - 1)))	//blink green if going to receive from current navigate-to device
+					{
+						led_green_on();
+					}
+				}
     		}
-        }
-        else
-        {
-    		timer1_stop_reload();
-    		time_slot_timer_ovf = 0;
-    		time_slot = 0;
 
-    		if (p_gps_num->status == GPS_DATA_VALID)
-        	{
-        		main_flags.frame_ended = 1; //if gps is valid then calculate rel pos for all devices
-        	}
-        	else
-        	{
-        		main_flags.frame_ended = 0;
-        	}
-        }
-
+    	    //after all do a re-calculate all
+    	    main_flags.process_all = 1;
+    	}
     }
 }