+++ CubeIDE/Code/src/timer.c
@@ -16,8 +16,6 @@ void systick_start(void);
 void systick_set_100ms(void);
 void systick_set_1000ms(void);
 void timer1_init(void);
-void timer1_clock_disable(void);
-void timer1_clock_enable(void);
 void timer2_init(void);
 void timer2_start(void);
 void timer3_init(void);
@@ -26,6 +24,7 @@ void timer4_init(void);
 
 
 uint8_t sound_enabled = 1; //status of the beep sound notification
+uint8_t timer1_interval_type = 0; // 1 - long; 2 - short
 
 
 
@@ -47,10 +46,11 @@ void make_a_beep(void)
 	if (sound_enabled)
 	{
 		timer2_start();		//pwm
-		systick_set_100ms();
-		systick_start();	//gating
-		led_red_on();		//todo: LED should blink even if sound is off
 	}
+
+	systick_set_100ms();
+	systick_start();	//gating
+	led_red_on();
 }
 
 
@@ -60,10 +60,11 @@ void make_a_long_beep(void)
 	if (sound_enabled)
 	{
 		timer2_start();		//pwm
-		systick_set_1000ms();
-		systick_start();	//gating
-		led_red_on();
 	}
+
+	systick_set_1000ms();
+	systick_start();	//gating
+	led_red_on();
 }
 
 
@@ -145,37 +146,40 @@ void timer1_init(void)
 {
     RCC->APB2ENR |= RCC_APB2ENR_TIM1EN;     //enable timer 1 clock
     TIM1->PSC = (uint16_t)2999;            	// 3MHz/(2999+1)=1kHz
-    TIM1->ARR = (uint16_t)49;               // 1kHz/(49+1)=20Hz (50ms)
+//    TIM1->ARR = (uint16_t)499;              // 1kHz/(499+1)=2Hz (500ms)
     TIM1->CR1 |= TIM_CR1_URS;               //only overflow generates interrupt
     TIM1->EGR = TIM_EGR_UG;                 //software update generation
     TIM1->SR &= ~TIM_SR_UIF;                //clear update interrupt
     TIM1->DIER |= TIM_DIER_UIE;             //update interrupt enable
 
     NVIC_EnableIRQ(TIM1_UP_IRQn);           //enable interrupt
-    timer1_clock_disable();
 }
 
 
 
-void timer1_clock_disable(void)
+//Timer1 start 800 ms
+void timer1_start_800ms(void)
 {
-	BIT_BAND_PERI(RCC->APB2ENR, RCC_APB2ENR_TIM1EN) = 0;
-}
-
-
-
-void timer1_clock_enable(void)
-{
-	BIT_BAND_PERI(RCC->APB2ENR, RCC_APB2ENR_TIM1EN) = 1;
-}
-
-
-
-//Timer1 start
-void timer1_start(void)
-{
-	timer1_clock_enable();
+	TIM1->ARR = (uint16_t)799;              // 1kHz/(799+1)=1.25Hz (800ms)
     TIM1->CR1 |= TIM_CR1_CEN;               //enable counter
+    timer1_interval_type = 1; //long
+}
+
+
+
+//Timer1 start 100 ms
+void timer1_start_100ms(void)
+{
+	TIM1->ARR = (uint16_t)99;              // 1kHz/(99+1)=10Hz (100ms)
+    TIM1->CR1 |= TIM_CR1_CEN;               //enable counter
+    timer1_interval_type = 2; //short
+}
+
+
+
+uint8_t timer1_get_intrvl_type(void)
+{
+	return timer1_interval_type;
 }
 
 
@@ -185,7 +189,6 @@ void timer1_stop_reload(void)
 {
     TIM1->CR1 &= ~TIM_CR1_CEN;              //disable counter
     TIM1->EGR = TIM_EGR_UG;                 //software update generation
-    timer1_clock_disable();
 }